+++ src/Release.php
@@ -16,7 +16,7 @@ class Release
     /**
      * Version of the SDK
      */
-    const VERSION = '1.8.2051';
+    const VERSION = '1.8.2052';
 
     /**
      * @param Event $event