+++ src/ResourceSharing/ResourceSharing.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace AlibabaCloud\ResourceSharing;
+
+use AlibabaCloud\Client\Resolver\VersionResolver;
+
+/**
+ * @method static V20200110\ResourceSharingApiResolver v20200110()
+ */
+class ResourceSharing extends VersionResolver
+{
+}